# mall-server/apifox_openapi_3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cosplay Mall OpenAPI 3.0 测试集",
    "description": "基于 Swagger 2.0 自动生成的 Apifox 优化版接口文档，包含所有控制器的代表性接口，预设 Token 提取和变量引用。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9191",
      "description": "本地开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "token"
      }
    }
  },
  "paths": {
...
    "/api/address/{userId}": {
      "get": {
        "tags": ["address-controller"],
        "summary": "根据用户ID查询收货地址",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "integer", "default": "{{currentUserId}}" } },
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/address": {
      "post": {
        "tags": ["address-controller"],
        "summary": "新增收货地址",
        "parameters": [
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "integer", "example": "{{currentUserId}}" },
                  "linkUser": { "type": "string", "example": "测试员" },
                  "linkPhone": { "type": "string", "example": "13800138000" },
                  "address": { "type": "string", "example": "上海市浦东新区" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/afterSale/mine/page": {
      "get": {
        "tags": ["after-sale-controller"],
        "summary": "查询我的售后单分页",
        "parameters": [
          { "name": "pageNum", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/afterSale": {
      "post": {
        "tags": ["after-sale-controller"],
        "summary": "申请售后",
        "parameters": [
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderNo": { "type": "string", "example": "ORDER123" },
                  "reason": { "type": "string", "example": "商品破损" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/comment/good/{goodId}": {
      "get": {
        "tags": ["comment-controller"],
        "summary": "查询商品评价列表",
        "parameters": [
          { "name": "goodId", "in": "path", "required": true, "schema": { "type": "integer", "example": 1 } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/comment": {
      "post": {
        "tags": ["comment-controller"],
        "summary": "发表评价",
        "parameters": [
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "goodId": { "type": "integer", "example": 1 },
                  "content": { "type": "string", "example": "质量很好！" },
                  "star": { "type": "integer", "example": 5 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/good/page": {
      "get": {
        "tags": ["good-controller"],
        "summary": "商品列表分页查询",
        "parameters": [
          { "name": "pageNum", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "searchText", "in": "query", "schema": { "type": "string", "example": "礼服" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/good": {
      "post": {
        "tags": ["good-controller"],
        "summary": "新增/保存商品",
        "parameters": [
          { "name": "token", "in": "header", "required": true, "schema": { "type": "string", "default": "{{token}}" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "新品服装" },
                  "price": { "type": "number", "example": 299.0 },
                  "categoryId": { "type": "integer", "example": 1 },
                  "imgs": { "type": "string", "example": "img_url" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}